moved tests from unit dir to feature
